Deprecated kombu RPC driver
In the effort to keep mistral code running, we decided to get rid of RPC
kombu driver.
This driver does not bring something more that oslo.messaging.
It's also not the default and may not be used very often on mistral
deployments since the configuration reference was not even giving any
info about it.

This will simplify mistral code and will ease our objective to get rid
of eventlet in the future.

Change-Id: Ie10d0b8b94d4149e3e626a0146d64812c85704a0
Signed-off-by: Arnaud M <arnaud.morin@gmail.com>