Replace eventlet timeout in default_executor
Running action was done within an eventlet timeout (with statement),
which rely on greenthread to kill threads after a certain amount of
time (timeout).

Switching this logic to threading is not easy, because we can't really
kill threads correctly.

So, I decided to move the action.run() logic inside a thread and wait
timeout seconds for the thread to finish.
If the thread is not finished after that period, mistral will print a
LOG warning to let operator know that an action timed out but the thread
is still ongoing in the background.
During that time, the action will stay in RUNNING.
After the thread eventually finish, an exception will be raised.

On the opposite, the workflow and task will be in ERROR thanks to an
already existing logic in task scheduling (task will timeout, ending in
ERROR and pushing the workflow to ERROR even if action is still RUNNING).

Finally, the logic we had with eventlet before will still be the same
most of the time because I don't expect the thread to take infinite time.
If it happen, the worker will be blocked, but it will be visible from
logs, and I expect operator will understand what's going on in this
action, and eventually fix the action.
Note that, no default mistral action will take infinite time, so we
should be safe on that part.

Change-Id: I0d84c1fcf23fac9f54031b2d0890045aece233a1
Signed-off-by: Arnaud M <arnaud.morin@gmail.com>