Simplify the TimeoutThreadWithException class
I introduce the class TimeoutThreadWithException recently to help early
fail in some of the tests if it took too long.
While the initial implementation was working fine, it's a little bit
over engineered and the logic is not fully correct.

Let's adopt this new version with only a threading timer that set an
exeption at the end.

Change-Id: I2a0320747dd2a6061be2a9f47c0f805d7e5cb4fb
Signed-off-by: Arnaud M <arnaud.morin@gmail.com>