Add allowlist and denylist for action providers and actions
This adds allowlist and denylist ListOpt config options for
filtering what action providers is loaded

Also adds allowlist and denylist ListOpt config options for
filtering what actions is loaded by the legacy action provider.

This is used in a security context where we need to restrict
a environment to only load actions using the legacy action
provider and then filter what actions is explicitly allowed
or not allowed to be used.

Change-Id: I57cd468cba227cc8cc7960e84e9a111217e80cad